backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
alembic
pydantic-settings
python-multipart
redis
